move tests to separate file
